Update README.md
Another tech stack added to my Batman gadget belt! Here is my dockerized and deployed MERN stack project where I've replaced the R with Next.js. After getting criticism about the design for my Spring full stack project I took it as a task to prove to myself I was capable of producing good design. I decided on making a food delivery website as my training institute, TAP Academy, had a trend of teaching their students to build Java full stack web applications with the same theme in mind. I took my unformulated plan to Figma to make sense of how to go about it. I built the design purely from my own taste, the color scheme, component positioning and overall look of the web application. By this time the plan didn't seem so far fetched and it seemed possible to express myself through my projects, and what's more is that I was genuinely having fun making it and bringing it all together. For so long I thought it always had to be bitter work, but here I was actually enjoying what I do? Wow. I started building the project's front end on my local system using React as soon as I was done with the designs for the base pages. I had fun styling each component and the idea of each component possessing it's own stylesheet that didn't conflict with the other stylesheets. Usually I had problems from not being able to maintain my vision on code because I my mind was tired from making the selectors work with each other. I started learning the framework and actively applying it on my projects because my roommate had convinced me that it was the best for design. While it falls short in many areas, he wasn't wrong. It was definitely easier for me to bring my designs to life. It was between this phase that a friend had approached me to make a landing page for his business idea. Soul Notes, my previous project was built on the same framework and deployed in between my development of this project's front end development. It took some time to return, but once I did, I was unsatisfied with continuing my build on the same framework I had showcased. This was when I got introduced to Next.js. I would later find out using this React framework for front end was like using a flamethrower to light the fireplace. However, I had already decided that I was gonna push my comfort boundaries and so, with a working front end I started on the backend. I was set on building a MERN stack project as it was a buzzword that was thrown around during my college time and many people would fear not knowing enough when talking to someone that dropped "MERN stack" in a conversation. MongoDB was not hard to work with due it's unstructured query language, although the actual preparation of the custom database did take some time as I had to manually add the entries, but once the hard stuff is over the rest becomes not that hard. And after preparing all the tables with data that I wanted to work with, I had reached the core of the MERN stack, the Node.js part. Actively learning and putting my knowledge to the test was definitely helping me to grasp the concepts faster. I will admit that it was confusing, but once I got the hang of it, I was able to write enough code that I needed for my project. And besides, who doesn't love to work with APIs. I brought all the features together locally and was ready to post it then, but I had been attending interviews, applying for jobs and had fortunately gotten placed at a company at the time of almost completing my project, I had some creases I wanted to iron out before pushing it to deployment so i decided it's fine to delay it a bit so as to get my career initialization done. The company I work in had tasked me to familiarize myself with Docker, Inc, the tool and Kubernetes. And I saw an alignment with what I wanted to do. So this project has served as a proof of my training in Docker. While I haven't mastered it, I had learnt enough to implement what I needed in my project. The docker desktop had some straightforward learning material which i put to good use. After a few trial and errors, I had successfully dockerized my MENN stack project and it was ready for deployment. I used Vercel for deploying my front end and Render to dpeloy my backend. I had tried to fix the design so that it looks good through all devices, but reverted to the vanilla build when it got out hand. I do plan to keep working on it, but I just wanted to come and share my journey and the project itself. Feel free to take a look at it, I'd advise checking it out on your desktops as it's still under development for the responsive design.